cdo: move expansion logic of yaml flags to xlate-yaml
Move the 'yaml to dts' flags expansion logic to xlate-yaml assist.
