+++ assists/cdo.py
@@ -51,144 +51,6 @@ def is_compat(node, compat_id):
     return ""
 
 
-def do_expand_cdo_flags(tgt_node):
-    if 'xilinx,subsystem-v1' in tgt_node.propval("compatible"):
-        return True
-    return False
-
-
-def expand_cdo_flags(tgt_node):
-    flags_names = []
-    flags = []
-    # find default flags
-    for n in tgt_node.subnodes():
-        if n.depth == tgt_node.depth + 2 \
-                and n.abs_path == tgt_node.abs_path + "/flags/default":
-            default_flags = n
-            break
-
-    # for each flag reference
-    # update flags' bits along with using default
-    for n in tgt_node.subnodes():
-        if n.depth == tgt_node.depth + 2:
-            #    set all bits according to what is provided
-            flags.extend(expand_cdo_flags_bits(n, default_flags))
-            flags_names.append(n.name)
-
-    flags_cells = 4
-    return [flags_names, flags, flags_cells]
-
-
-def expand_cdo_flags_bits_first_word(flags_node, ref_flags,
-                                     default_flags_node):
-    first_keywords = {
-        'allow-secure': 2,
-        'read-only': 4,
-        'requested': 6
-    }
-    allow_sec = False
-    read_only = False
-    for key in (list(first_keywords.keys())):
-
-        # determine which node to read flags info from for this bit
-        in_flags = flags_node.propval(key) != [''] or \
-            key in flags_node.__props__.keys()
-        in_default = key in default_flags_node.__props__.keys()
-        node = None
-        if in_flags == True:
-            node = flags_node
-        elif in_flags == False and in_default == True:
-            node = default_flags_node
-        else:
-            continue
-
-        if node.propval(key) != [''] or key in node.__props__.keys():
-            if key == 'allow-secure' or key == 'read-only':
-                # if true then set to low, as per spec: "0: secure master only."
-                # write should be low if read-only
-                ref_flags[0] = ref_flags[0] & ~(0x1 << first_keywords[key])
-                if key == 'allow-secure':
-                    allow_sec = True
-                    continue
-                elif key == 'read-only':
-                    read_only = True
-                    continue
-
-            ref_flags[0] |= (0x1 << first_keywords[key])
-
-    # 1st word read policy always allowed
-    #ref_flags[0] = ref_flags[0] | (0x1 << 3)
-
-    # 1st word: if allow-secure is false, then raise corresponding bit
-    if allow_sec == False:
-        ref_flags[0] |= (0x1 << first_keywords['allow-secure'])
-    # 1st word: if read only  is false, then raise corresponding bit
-    if read_only == False:
-        ref_flags[0] |= (0x1 << first_keywords['read-only'])
-
-    # 1st word time share
-    if flags_node.propval('timeshare') != ['']:
-        ref_flags[0] |= 0x3
-
-
-def expand_cdo_flags_bits_third_word(flags_node, ref_flags,
-                                     default_flags_node):
-    third_keywords = {
-        'access': 0,
-        'context': 1,
-        'wakeup': 2,
-        'unusable': 3,
-        'requested-secure': 4,
-        'coherent': 5,
-        'virtualized': 6
-    }
-
-    for key in third_keywords.keys():
-        # determine which node to read flags info from for this bit
-        in_flags = flags_node.propval(key) != [''] \
-            or key in flags_node.__props__.keys()
-        in_default = key in default_flags_node.__props__.keys()
-
-        node = None
-        if in_flags == True:
-            node = flags_node
-        elif in_flags == False and in_default == True:
-            node = default_flags_node
-        else:
-            continue
-
-        if node.propval(key) != [''] or key in node.__props__.keys():
-
-            # this can only be set if prealloc is set too
-            if key == 'requested-secure':
-                if (ref_flags[0] & (0x1 << first_keywords['requested'])) == 0:
-                    continue
-
-            ref_flags[2] |= 0x1 << third_keywords[key]
-
-
-# given flags node, return int that is bitmask of relevant bits
-def expand_cdo_flags_bits(flags_node, default_flags_node):
-    ref_flags = [0x0, 0x0, 0x0, 0x0]
-
-    expand_cdo_flags_bits_first_word(flags_node, ref_flags, default_flags_node)
-
-    # 2nd word is derived
-    ref_flags[1] = 0xfffff
-
-    expand_cdo_flags_bits_third_word(flags_node, ref_flags, default_flags_node)
-
-    # 4th word
-    qos = 0x0
-    if flags_node.propval('qos') != [''] and isinstance(qos, int):
-        qos = flags_node.propval('qos')
-    elif default_flags_node.propval('qos') != [''] and isinstance(qos, int):
-        qos = default_flags_node.propval('qos')
-    ref_flags[3] = qos
-
-    return ref_flags
-
-
 class Device():
     def __init__(self, nodes, node_id, flags):
         self.nodes = nodes
